Add files via upload
Hello world!
I am here with this amazing dashboard which contains top performers as per sales dataset from the automobile sector, also trying to focus on data points that act as relevant proof when we plan to choose a market for our product and its placement. 
Hope you found it meaningful with the provided details.